moved highlight colors into separate file and renamed .hpp files to .h
