+++ Set-Finder/FrameProcessorWrapper.h
