+++ Set-Finder/Set-Finder-Bridging-Header.h
@@ -2,4 +2,4 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
-#import "FrameProcessorWrapper.hpp"
+#import "FrameProcessorWrapper.h"
