+++ Set-Finder/cpp/include/SetGame.h
