+++ Set-Finder/cpp/src/FrameProcessor.cpp
@@ -5,8 +5,9 @@
 //  Created by JD del Alamo on 7/19/23.
 //
 
-#include "FrameProcessor.hpp"
-#include "SetGame.hpp"
+#include "FrameProcessor.h"
+#include "SetGame.h"
+#include "HighlightColors.h"
 
 const float MIN_CARD_AREA_PERCENTAGE = 0.007;
 
@@ -53,72 +54,6 @@ const int OPEN_SHADING_CONTRAST_THRESHOLD = 25;
 const int STRIPED_SHADING_CONTRAST_THRESHOLD = 125;
 
 const float HIGHLIGHT_SCALE_FACTOR = 0.15;
-
-const std::vector<cv::Scalar> SET_HIGHLIGHT_COLORS = {
-   cv::Scalar(190, 94, 232),
-   cv::Scalar(255, 155, 0),
-   cv::Scalar(103, 177, 255),
-   cv::Scalar(210, 255, 165),
-   cv::Scalar(64, 87, 167),
-   cv::Scalar(78, 173, 95),
-   cv::Scalar(130, 104, 107),
-   cv::Scalar(57, 95, 0),
-   cv::Scalar(65, 110, 255),
-   cv::Scalar(120, 255, 0),
-   cv::Scalar(255, 0, 181),
-   cv::Scalar(177, 68, 117),
-   cv::Scalar(82, 255, 152),
-   cv::Scalar(156, 143, 0),
-   cv::Scalar(0, 14, 98),
-   cv::Scalar(2, 229, 255),
-   cv::Scalar(198, 255, 0),
-   cv::Scalar(116, 255, 222),
-   cv::Scalar(44, 0, 67),
-   cv::Scalar(0, 136, 187),
-   cv::Scalar(232, 138, 150),
-   cv::Scalar(112, 153, 190),
-   cv::Scalar(203, 208, 145),
-   cv::Scalar(161, 76, 14),
-   cv::Scalar(49, 130, 120),
-   cv::Scalar(254, 111, 229),
-   cv::Scalar(84, 71, 0),
-   cv::Scalar(255, 208, 1),
-   cv::Scalar(163, 116, 255),
-   cv::Scalar(159, 140, 194),
-   cv::Scalar(68, 21, 0),
-   cv::Scalar(142, 0, 158),
-   cv::Scalar(23, 185, 0),
-   cv::Scalar(147, 211, 189),
-   cv::Scalar(0, 52, 38),
-   cv::Scalar(255, 198, 189),
-   cv::Scalar(59, 61, 104),
-   cv::Scalar(126, 174, 0),
-   cv::Scalar(0, 36, 164),
-   cv::Scalar(86, 0, 255),
-   cv::Scalar(0, 169, 133),
-   cv::Scalar(210, 45, 126),
-   cv::Scalar(130, 71, 122),
-   cv::Scalar(0, 137, 254),
-   cv::Scalar(157, 2, 255),
-   cv::Scalar(108, 130, 106),
-   cv::Scalar(126, 147, 255),
-   cv::Scalar(0, 255, 213),
-   cv::Scalar(255, 118, 0),
-   cv::Scalar(146, 251, 144),
-   cv::Scalar(0, 77, 119),
-   cv::Scalar(232, 238, 255),
-   cv::Scalar(246, 0, 255),
-   cv::Scalar(181, 125, 0),
-   cv::Scalar(58, 0, 149),
-   cv::Scalar(103, 0, 1),
-   cv::Scalar(1, 100, 0),
-   cv::Scalar(102, 219, 255),
-   cv::Scalar(254, 166, 255),
-   cv::Scalar(254, 255, 1),
-   cv::Scalar(0, 0, 255),
-   cv::Scalar(255, 0, 0),
-   cv::Scalar(0, 255, 0)
-};
 
 void
 FrameProcessor::process(cv::Mat& frame)
